test_common.h: modify to use double instead of float in round_to_nearest_even function

+++ include/test_common.h
@@ -94,10 +94,10 @@ Halide::Runtime::Buffer<T> mk_const_buffer(const std::vector<int32_t>& extents, 
 }
 
 template<typename T>
-T round_to_nearest_even(float v)
+T round_to_nearest_even(double v)
 {
-    float i;
-    float f = modff(v, &i);
+    double i;
+    double f = modf(v, &i);
     if (0.5 == f || -0.5 == f) {
         if (0 == static_cast<int64_t>(i) % 2) {
             return static_cast<T>(i);
